all conversations checked
