+++ internal/driver/readingchecker.go
@@ -1,6 +1,5 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
 //
-// Copyright (C) 2019-2021 IOTech Ltd
 // Copyright (C) 2023 YIQISOFT
 //
 // SPDX-License-Identifier: Apache-2.0
