+++ internal/driver/utils.go
@@ -1,7 +1,5 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
 //
-// Copyright (C) 2021 IOTech Ltd
-// Copyright (C) 2022 Schneider Electric
 // Copyright (C) 2023 YIQISOFT
 //
 // SPDX-License-Identifier: Apache-2.0
